Update fft.py
I found a bug in FFT when it runs for a large number of points; it worked before because the number of points in the test case was limited to 4.  I did some updates but it still has the bug and I believe the bug is because of the algorithm provided in the paper. I provided the paper and a thesis cited the paper in the following:
http://www.cs.technion.ac.il/users/wwwb/cgi-bin/tr-get.cgi/2016/MSC/MSC-2016-15.pdf, page 22
http://www.math.clemson.edu/~sgao/papers/GM10.pdf, page 3
the bug is that we need to construct a list G, and later we use G. Where G is used, there is an assumption which is the size of G in 2^{m-1}, while where we construct G, the size of G is m. Since before I tested the algorithm for number of points 4 and m = 2, it worked, but obviously the algorithm has an issue if m is bigger than 2!
I returns back the polynomial interpolation in APR to the original one.
Please share your thoughts in this case.